Call sit-for using new calling convention
Since Emacs 22.1 the function is advertised as only supporting
one or two arguments, but actually it still supports up to three.
Stop calling it with three arguments here, which means that we
drop support for older Emacsen.

Do longer claim what this works with ancient Emacsen that we don't
actually test.  Set Package-Requires to depend on Emacs 25, even
though this might still work with older releases.